pyproject.toml
[project]
name = "langchain-streamlit-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
